requirements.txt
instagrapi
google-cloud-secret-manager==2.15.1
